refactor: Move dates and files modules to support folder
