+++ src/program.js
@@ -2,7 +2,7 @@ const { Command, Option } = require("commander");
 const program = new Command();
 
 const { setDatesInFolder } = require("./assistant/runner");
-const { toAbsolute } = require("./files/utils");
+const { toAbsolute } = require("./support/files");
 const { setLevel } = require("./support/tracer");
 
 const logOption = new Option("-l, --log <logLevel>", "Log level").choices([
